Add transitclock.core.disableAssignmentGrab config param
Allows disabling assignment grabbing separate from the decision whether to have exclusive block assignments. If `transitclock.core.disableAssignmentGrab` is set to true (false by default), and it typically would have grabbed the assignment from an existing vehicle on that block, it will not do so. May be useful in situations where there's a lot of "thrashing" back and forth between vehicles fighting over the same assignment, and it is better to leave the assignment with the first vehicle that matched to it.