Update output GTFS Realtime feeds to version 2.0
The main difference in GTFS Realtime v2.0 is what fields are considered required or conditionally required. From what I have seen so far, TheTransitClock already generates all of the correct fields, so this is just a matter of communicating what version of the spec its output feeds are meant to conform to.

For more information see:

* https://github.com/google/gtfs-realtime-bindings/issues/26#issuecomment-395087743
* https://medium.com/@sjbarbeau/whats-new-in-gtfs-realtime-v2-0-cd45e6a861e9