Fix bugs in AVL playback map
Fixes the following bugs:
1. date picker's initial value works, but after using the date picker, the format changes and breaks the subsequent request
2. you can select a route and it will show it on the map but not limit the query results to avl reports for that route
3. returned JSON is not camelCased so the tooltips are broken

The underlying cause of 3, at least for deployments using PostgreSQL, is that PostgreSQL flattens all table and column names to lowercase. The schema for TheTransitClock are specified with camelCased column names, which don't follow through to the actual database when they're initially run to create the tables. So the subsequent returned rows that are translated to JSON are all lowercase. This is kind of a kludgy workaround.