Add `transitclock.avl.doubleMapAllowedMismatchedStops` to allow giving some leeway when matching DoubleMap routes to GTFS routes
This value sets how many additional stops to allow in the DoubleMap route info after matching with a potential GTFS route. Suggest starting at 0 and seeing if all routes match, but can set to a higher number to loosen how strict the matching is. If it's too high, will get multiple potential matches, which will result in not adding that route to the lookup.

Initial motivation was a GTFS static update where there was 1 old stop in the DoubleMap feeds (it had been removed in the latest version of the GTFS feed). This was causing a couple of routes to not match.